new clas deSerialization
